remove local composables package and move user row to users screen
