+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/presentation/users/UserScreen.kt
@@ -1,18 +1,22 @@
 package org.thechance.common.presentation.users
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import cafe.adriel.voyager.core.screen.Screen
@@ -23,11 +27,11 @@ import org.koin.core.component.inject
 import org.thechance.common.presentation.composables.BpDropdownMenu
 import org.thechance.common.presentation.composables.PermissionsFlowRow
 import org.thechance.common.presentation.composables.modifier.cursorHoverIconHand
+import org.thechance.common.presentation.composables.modifier.noRipple
 import org.thechance.common.presentation.composables.table.BpPager
 import org.thechance.common.presentation.composables.table.BpTable
 import org.thechance.common.presentation.composables.table.TotalItemsIndicator
 import org.thechance.common.presentation.uistate.UserScreenUiState
-import org.thechance.common.presentation.users.composables.UserRow
 
 
 object UserScreen : Screen, KoinComponent {
@@ -288,5 +292,77 @@ object UserScreen : Screen, KoinComponent {
                 onFilterPermissionClicked = onFilterPermissionClicked,
             )
         }
+    }
+
+    @OptIn(ExperimentalLayoutApi::class)
+    @Composable
+    fun RowScope.UserRow(
+        onClickEditUser: (UserScreenUiState.UserUiState) -> Unit,
+        position: Int,
+        user: UserScreenUiState.UserUiState,
+        firstColumnWeight: Float = 1f,
+        otherColumnsWeight: Float = 3f,
+    ) {
+        Text(
+            position.toString(),
+            style = Theme.typography.titleMedium.copy(color = Theme.colors.contentTertiary),
+            modifier = Modifier.weight(firstColumnWeight),
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis,
+        )
+
+        Row(Modifier.weight(otherColumnsWeight), verticalAlignment = Alignment.CenterVertically) {
+            Image(painter = painterResource("dummy_img.png"), contentDescription = null)
+            Text(
+                user.fullName,
+                style = Theme.typography.titleMedium.copy(color = Theme.colors.contentPrimary),
+                modifier = Modifier.padding(start = 16.dp),
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+        }
+
+        Text(
+            user.username,
+            style = Theme.typography.titleMedium.copy(color = Theme.colors.contentPrimary),
+            overflow = TextOverflow.Ellipsis,
+            modifier = Modifier.weight(otherColumnsWeight),
+            maxLines = 1,
+        )
+        Text(
+            user.email,
+            style = Theme.typography.titleMedium.copy(color = Theme.colors.contentPrimary),
+            overflow = TextOverflow.Ellipsis,
+            modifier = Modifier.weight(otherColumnsWeight),
+            maxLines = 1,
+        )
+        Text(
+            user.country,
+            style = Theme.typography.titleMedium.copy(color = Theme.colors.contentPrimary),
+            overflow = TextOverflow.Ellipsis,
+            modifier = Modifier.weight(otherColumnsWeight),
+            maxLines = 1,
+        )
+        FlowRow(
+            Modifier.weight(otherColumnsWeight),
+            horizontalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            user.permissions.forEach {
+                Icon(
+                    painter = painterResource(it.iconPath),
+                    contentDescription = null,
+                    tint = Theme.colors.contentPrimary.copy(alpha = 0.87f),
+                    modifier = Modifier.size(24.dp)
+                )
+            }
+        }
+
+        Image(
+            painter = painterResource("horizontal_dots.xml"),
+            contentDescription = null,
+            modifier = Modifier.noRipple { onClickEditUser(user) }
+                .weight(firstColumnWeight),
+            colorFilter = ColorFilter.tint(color = Theme.colors.contentPrimary)
+        )
     }
 }
